pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btc-analytics"
version = "0.1.0"
description = "Local-first Bitcoin parser and on-chain metrics engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
